+++ ts-src/client/extension.ts
@@ -16,34 +16,20 @@ export async function activate(context: ExtensionContext) {
   const commonOptions = {
     command,
     args,
-    options: {
-      env: {
-        ...process.env,
-        OCAMLRUNPARAM: "b",
-        MERLIN_LOG: "-",
-      },
-    },
+    options: { env: { ...process.env, OCAMLRUNPARAM: "b", MERLIN_LOG: "-" } },
   };
 
-  // If the extension is launched in debug mode then the debug server options are used
-  // Otherwise the run options are used
   const serverOptions: ServerOptions = {
     run: commonOptions,
     debug: commonOptions,
   };
 
-  // Options to control the language client
   const clientOptions: LanguageClientOptions = {
-    // Register the server for plain text documents
     documentSelector: [
       { scheme: "file", language: "ocaml" },
       { scheme: "file", language: "reason" },
       { scheme: "file", language: "ocamllex" },
     ],
-    synchronize: {
-      // Notify the server about file changes to '.clientrc files contained in the workspace
-      fileEvents: workspace.createFileSystemWatcher("**/.clientrc"),
-    },
   };
 
   const createClient = () => {
@@ -55,10 +41,8 @@ export async function activate(context: ExtensionContext) {
     );
   };
 
-  // Create the language client and start the client.
   client = createClient();
 
-  // Start the client. This will also launch the server
   client.start();
 
   commands.registerCommand("ocaml-language-server.restart", () => {