+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## Unreleased
+
+- Improve ocamllex syntax highlighting
+- Improve opam syntax highlighting
+
 ## 0.4.0
 
 - Add syntax highlighting and basic language support for ocamlyacc/menhir