+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Improve ocamllex syntax highlighting
 - Improve opam syntax highlighting
+- Fix bugs in ocaml and ocamllex syntax highlighting
 
 ## 0.4.0